+++ script/dev/commands.ts
@@ -94,6 +94,42 @@ const developerCommands: any = [
     },
   },
   {
+    name: "enter",
+    description: "[dungeon] Enter a dungeon",
+    help: "fight [enemy] [amount] - Enter a specified dungeon.<br>Example: enter vithail_dungeon",
+    isCheat: true,
+    list: [dungeons],
+    execute: (args: string[]) => {
+      const dungeon = args[0];
+      if (dungeon) {
+        const index = dungeons.findIndex((d: any) => d.id === dungeon);
+        if (index !== -1) {
+          dungeonController.enterDungeon(dungeons[index]);
+          devConsole.commandHistory.push(`Entered dungeon ${dungeon}`);
+        } else {
+          devConsole.commandHistory.push(`Dungeon "${dungeon}" not found`);
+        }
+      } else {
+        devConsole.commandHistory.push("Too few arguments, expected: enter [dungeon]");
+      }
+    },
+  },
+  {
+    name: "leave",
+    description: "Leave the current dungeon",
+    help: "leave - Leave the current dungeon.",
+    isCheat: true,
+    list: [],
+    execute: () => {
+      if (dungeonController.currentDungeon) {
+        devConsole.commandHistory.push(`Left dungeon ${dungeonController.currentDungeon.id}`);
+        dungeonController.leaveDungeon();
+      } else {
+        devConsole.commandHistory.push("Not in a dungeon");
+      }
+    },
+  },
+  {
     name: "fight",
     description: "[enemy] [amount] Fight an enemy",
     help: "fight [enemy] [amount] - Fight an enemy. Specify amount for group battle.<br>Example: fight skeleton 2",
@@ -112,10 +148,31 @@ const developerCommands: any = [
           devConsole.commandHistory.push(`Enemy "${enemy}" not found`);
         }
       } else {
-        devConsole.commandHistory.push("Too few arguments, expected: fight [enemy] [amount");
+        devConsole.commandHistory.push("Too few arguments, expected: fight [enemy] [amount]");
       }
     },
   },
+  {
+    name: "killall",
+    description: "Kill all enemies in the current fight.",
+    help: "killall - Kill all enemies in the current fight.",
+    isCheat: true,
+    list: [],
+    execute: () => {
+      const _enemies = combat.getLivingEnemies();
+      if (_enemies.length > 0) {
+        _enemies.forEach((enemy: Enemy) => {
+          enemy.stats.hp = 0;
+          update();
+          enemy.die();
+        });
+        devConsole.commandHistory.push("All enemies killed");
+      } else {
+        devConsole.commandHistory.push("No enemies to kill");
+      }
+    },
+  },
+
   {
     name: "item",
     description: "[item] [amount] Adds specified quantity of items to your inventory",