+++ script/dev/dev.ts
@@ -16,7 +16,11 @@ const DEVTOOLS: DevTools = {
 
 if (localStorage.getItem("devtools") === "true") DEVTOOLS.ENABLED = true;
 
-const gameVersion = "0.1.1";
+const gameVersion = (0.12).toFixed(2);
+const gameVersionText = (ver: string) => {
+  if (parseFloat(ver) < 0.1) return "Old";
+  return `${parseInt(ver)}.${(parseFloat(ver) * 10) % 10}`;
+};
 
 // Skip main menu
 if (DEVTOOLS.ENABLED) {
@@ -24,11 +28,11 @@ if (DEVTOOLS.ENABLED) {
   mainMenuElement.classList.add("no-display");
   lobby.current_view = "perks";
   createLobby();
-  //dungeonController.enterDungeon(dungeons[0]);
+  //dungeonController.enterDungeon(dungeons[1]);
 }
 
 function convertRemToPixels(rem: number) {
   return rem * parseFloat(getComputedStyle(document.documentElement).fontSize);
 }
 
-document.title = `Pohjola III - v${gameVersion}`;
+document.title = `Pohjola III - v${gameVersionText(gameVersion)}`;