+++ script/dungeon/dungeon.ts
@@ -375,6 +375,7 @@ class DungeonController {
     dungeonContent.innerHTML = "";
     this.currentDungeon = new Dungeon(JSON.parse(JSON.stringify(dungeon)));
     this.currentRoom = this.currentDungeon.rooms[0];
+    this.canMove = true;
     buildDungeon();
   }