+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+# Version 0.1.2 ~ 2023.3.14
+
+## New features & additions
+- Added new dungeon: `Abandoned Fort`
+  * Unlocked after defeating the troll.
+  * Fight boss "Orc Chieftain" to complete.
+- Added orc related items.
+  * Orc chief armor set.
+  * Orc berserker axes.
+- Added "killall" command to the game.
+- Added "enter" command to the game.
+- Added "leave" command to the game.
+
+## Bug fixes
+- Fixed text overflowing from notifications.
+- Fixed stun resistance doing nothing.
+
 # Version 0.1.1 ~ 2023.3.12
 
 ## New features & additions
@@ -6,7 +23,7 @@
 - Added the "ranger" armor set. 
   * This armor set is a bit weaker than the "plate" armor set, but it has a chance to dodge attacks.
 - Added some missing localization.
-# Bug fixes
+## Bug fixes
 - Fixed being able to upgrade the first perk, despite it having no levels. 
   * This happened because the max level check was done after a requirement check  
   which the first perk passes by default.